Make possible to re open a stream
The user case: within a method, 1) entire file is read into local
variable, 2) process the contents, 3) write to the file, replace it's
contents with the processed contents. In this case the stream has to be
opened twice within a method, first time with read file access and
second time with write.

Method ReOpenAs disposes the current stream and open it again with new
parameters.